+++ src/addons/stories/RunList.stories.svelte
@@ -2,21 +2,23 @@
   import { Meta, Story, Template } from "@storybook/addon-svelte-csf";
 
   import RunList from "../runs/RunList.svelte";
-  import { results as events } from "../fixtures/event-list.json";
   import { results as runs } from "../fixtures/run-list.json";
 
-  const args = { visible: true, events, runs };
+  const args = { runs };
 </script>
 
+<style></style>
+
 <Meta
-  title="Add-Ons / Run list"
-  tags={["autodocs"]}
-  parameters={{ layout: "fullscreen" }}
+  title="Add-Ons / Runs"
   component={RunList}
+  parameters={{ layout: "centered" }}
 />
 
 <Template let:args>
-  <RunList {...args} />
+  <div class="container">
+    <RunList {...args} />
+  </div>
 </Template>
 
-<Story name="Default" {args} />
+<Story name="Single" {args} />
